Project Reramp Prepwork
Moved all old files to an "Old Package", slowly begining to draw them out and reorganize them into the new Generation Scheme. With it being an "Template ActionList" based System, and each Generator keeping track of the rooms it generates, it should be easier to manage and give more flexability while remaining cleaner. I also need to find a new "Naming Scheme", as "GenerationMap" and "GenerationMapGenerator" does not work, while "Dungeon" can be confussed with the "Live Dungeon" object which should keep track of rooms/events/treasure/mobs and stuff like that once it hits the plugin