Hallway Subdividing rooms working
Subdividing Rooms working perfectly! With the Blending, Randomized Schematics, it looks great!

A bug resulting in "Dead" Hallways not connected to BigRooms due to code not forcing connections ( Instead opting to allow failed connections ). This has been resolved and it should force connections when able.

Next step is to get bigrooms to generate